integrations/requirements.txt
# Gemini Function Calling integration
google-generativeai>=0.3.0
httpx>=0.26.0
